expose proxy url config on the endpoint
